+++ src/Core/Credentials/Scope.php
@@ -18,6 +18,7 @@ class Scope
      */
     protected array $availableScope = [
         'bizproc',
+        'biconnector',
         'calendar',
         'call',
         'cashbox',
@@ -98,5 +99,13 @@ class Scope
     public function getScopeCodes(): array
     {
         return $this->currentScope;
+    }
+
+    /**
+     * @throws \Bitrix24\SDK\Core\Exceptions\UnknownScopeCodeException
+     */
+    public static function initFromString(string $scope): self
+    {
+        return new self(str_replace(' ', '', explode(',', $scope)));
     }
 }